refactor: split suggestion engine into separate generators
Extract `generateCatchSuggestions`, `generateGiftAndTradeSuggestions`,
`generateBreedingSuggestions`, and `generateEvolutionSuggestions` out
of `suggestionEngine.ts` into individual files inside `src/engine/assistant/generators/`.

Extract shared types and constants into their own module files to prevent circular dependencies.
