+++ .jules/sculptor.md
@@ -13,3 +13,8 @@ Extracted complex inline IIFE math calculations and tailwind styling logic for c
 - AI Readability Impact: Deeply nested inline IIFEs within the return statement obscure the render hierarchy, increase parsing complexity, and make targeted AST modifications significantly harder. Moving computational logic out of the JSX and replacing it with clean conditional renders (`{condition && ...}`) heavily improves component readability and structural purity.
 - Extracted complex inline `Object.entries(...reduce(...))` catch method grouping logic from `AssistantSuggestionCard.tsx` into a cleanly separated `useMemo` hook named `catchMethods`.
 - AI Readability Impact: Deeply nested, convoluted object reductions mapped inline within the return statement obscure the render hierarchy and increase cognitive overhead. Moving complex data manipulation logic into isolated `useMemo` hooks significantly improves readability, reduces visual noise in the JSX, and makes subsequent structural modifications much easier to reason about.
+
+## Structural Learnings
+- **Monolithic Engine Refactoring**: Successfully refactored `src/engine/assistant/suggestionEngine.ts` by splitting out giant logic functions (`generateCatchSuggestions`, `generateGiftAndTradeSuggestions`, `generateBreedingSuggestions`, `generateEvolutionSuggestions`) into dedicated module files under `src/engine/assistant/generators/`. This greatly reduces cognitive load when modifying suggestion logic and makes the codebase easier for AI agents to process contextually. Shared types and constants were moved to `suggestionEngineTypes.ts` and `constants.ts` to prevent circular dependencies.
+- **Biome Strictness**: Biome linting correctly flagged the need to re-sort imports after moving everything out. Using `pnpm check:fix` effectively applies auto-fixes without manually resolving line-by-line diffs.
+- **Vitest Browser Dependency**: Tests may unexpectedly fail with `browserType.launch: Executable doesn't exist` when `pnpm test` triggers browser tests on a clean image without playwright binaries downloaded. The resolution is to always run `pnpm exec playwright install` if this test error occurs.
