+++ src/engine/assistant/__tests__/suggestionEngine.fix.test.ts
@@ -2,8 +2,8 @@ import { describe, expect, it } from 'vitest';
 import type { PokemonMetadata } from '../../../db/schema';
 import type { SaveData } from '../../saveParser/index';
 import { gen1Strategy } from '../strategies/gen1Strategy';
-import type { AssistantApiData } from '../suggestionEngine';
 import { generateSuggestions } from '../suggestionEngine';
+import type { AssistantApiData } from '../suggestionEngineTypes';
 
 describe('suggestionEngine - Redundancy Fix Verification', () => {
   const mockSaveData = {
