+++ src/engine/assistant/__tests__/test-coverage.test.ts
@@ -1,8 +1,8 @@
 import { expect, test } from 'vitest';
 import type { PokemonInstance, SaveData } from '../../saveParser/index';
 import { gen1Strategy } from '../strategies/gen1Strategy';
-import type { AssistantApiData } from '../suggestionEngine';
 import { generateSuggestions } from '../suggestionEngine';
+import type { AssistantApiData } from '../suggestionEngineTypes';
 
 test('coverage for suggestionEngine new lines', () => {
   const mockSaveData: SaveData = {
