+++ src/engine/assistant/constants.ts
@@ -1,0 +1,12 @@
+import { ENCOUNTER_METHOD } from '../../db/schema';
+
+export const METHOD_NAMES: Record<number, string> = {
+  [ENCOUNTER_METHOD.WALK]: 'walk',
+  [ENCOUNTER_METHOD.SURF]: 'surf',
+  [ENCOUNTER_METHOD.OLD_ROD]: 'old-rod',
+  [ENCOUNTER_METHOD.GOOD_ROD]: 'good-rod',
+  [ENCOUNTER_METHOD.SUPER_ROD]: 'super-rod',
+  [ENCOUNTER_METHOD.GIFT]: 'gift',
+  [ENCOUNTER_METHOD.ROCK_SMASH]: 'rock-smash',
+  [ENCOUNTER_METHOD.HEADBUTT]: 'headbutt',
+};
